+++ src/http/asyncRoute.ts
@@ -1,0 +1,9 @@
+import type { NextFunction, Request, RequestHandler, Response } from 'express';
+
+export type AsyncHandler = (req: Request, res: Response, next: NextFunction) => Promise<unknown>;
+
+export function asyncRoute(handler: AsyncHandler): RequestHandler {
+  return (req, res, next) => {
+    void Promise.resolve(handler(req, res, next)).catch(next);
+  };
+}